refactor(trusty-common): split lib.rs into focused submodules under 500 SLOC (closes #1108)
lib.rs was 764 SLOC (allowlisted). Split inline code into six new
focused submodules:

- port.rs: bind_with_auto_port TCP port-walking helper
- data_dir.rs: DATA_DIR_OVERRIDE_ENV, resolve_data_dir,
  sanitize_data_root, is_dir
- daemon_addr.rs: write_daemon_addr, read_daemon_addr,
  check_already_running
- health_probe.rs: probe_health HTTP health-probe
- tracing_init.rs: init_tracing, init_tracing_with_buffer,
  init_tracing_with_buffer_and_capture (feature-gated), maybe_disable_color
- openrouter_legacy.rs: ChatMessage, openrouter_chat (deprecated),
  openrouter_chat_stream (deprecated)

lib.rs is now a thin 62-SLOC re-export facade. All pre-split public
paths remain reachable as trusty_common::X. The allowlist entry for
lib.rs is removed (ratchet down from 95 → 94 entries).

cargo check --workspace: clean
cargo test -p trusty-common: 63 passed, 0 failed
cargo clippy -p trusty-common --all-targets -D warnings: clean
bash scripts/check_line_cap.sh: 94 allowlisted, 0 violations
